requirements.txt
Flask==3.0.0
numpy==1.21.2
python-dotenv==0.19.0
requests==2.26.0
